+++ LookupAnything/release-notes.md
@@ -5,6 +5,7 @@
 * Updated for Stardew Valley 1.5, including...
   * split-screen mode and UI scaling;
   * added island crystal cave puzzle lookup;
+  * added island field office lookup;
   * added island mermaid puzzle lookup;
   * added island shrine puzzle lookup.
 * Fixed some untranslated text.